refactor(xtoken): rename tools/x_token/_shared.py to utils.py
Addresses PR #2508 review feedback: rename the projection-prep helper
module from the underscore-prefixed _shared.py to utils.py and update its
importers (minimal_projection_generator, minimal_projection_via_multitoken,
sort_and_cut_projection_matrix), the build_projection_matrix.sh comment,
and the unit test.
