+++ tools/x_token/sort_and_cut_projection_matrix.py
@@ -18,7 +18,7 @@ import re
 import torch
 import tqdm
 
-from tools.x_token._shared import sinkhorn_one_dim
+from tools.x_token.utils import sinkhorn_one_dim
 
 
 def parse_arguments() -> argparse.Namespace:
